Extracts logging setup
