+++ tests/conftest.py
@@ -1,39 +1,15 @@
 import sys
 import threading
-from pathlib import Path
 from types import TracebackType
 from typing import Any
 
 import structlog
 
 from pnpq.events import Event
+from tests.logs import setup_log
 
 
-def find_project_dir(path: Path) -> Path:
-    if path.is_dir():
-        for contained_path in path.iterdir():
-            if contained_path.name == "pyproject.toml":
-                return path
-    return find_project_dir(path.parent)
-
-
-target_dir = find_project_dir(Path(__file__).resolve()).joinpath("target")
-target_dir.mkdir(exist_ok=True)
-
-structlog.configure(
-    processors=[
-        structlog.processors.add_log_level,
-        structlog.processors.StackInfoRenderer(),
-        structlog.dev.set_exc_info,
-        structlog.processors.dict_tracebacks,
-        structlog.processors.TimeStamper(fmt="iso", utc=True),
-        structlog.processors.JSONRenderer(),
-    ],
-    logger_factory=structlog.WriteLoggerFactory(
-        file=target_dir.joinpath("tests.log").open("a", encoding="utf-8")
-    ),
-)
-
+setup_log("tests")
 log = structlog.get_logger()
