+++ tests/logs.py
@@ -1,0 +1,32 @@
+from pathlib import Path
+
+import structlog
+
+def find_project_dir(path: Path) -> Path:
+    if path.is_dir():
+        for contained_path in path.iterdir():
+            if contained_path.name == "pyproject.toml":
+                return path
+    return find_project_dir(path.parent)
+
+
+target_dir = find_project_dir(Path(__file__).resolve()).joinpath("target")
+target_dir.mkdir(exist_ok=True)
+
+def setup_log(log_file_name):
+    target_dir = find_project_dir(Path(__file__).resolve()).joinpath("target")
+    target_dir.mkdir(exist_ok=True)
+
+    structlog.configure(
+        processors=[
+            structlog.processors.add_log_level,
+            structlog.processors.StackInfoRenderer(),
+            structlog.dev.set_exc_info,
+            structlog.processors.dict_tracebacks,
+            structlog.processors.TimeStamper(fmt="iso", utc=True),
+            structlog.processors.JSONRenderer(),
+        ],
+        logger_factory=structlog.WriteLoggerFactory(
+            file=target_dir.joinpath(f"{log_file_name}.log").open("a", encoding="utf-8")
+        ),
+    )
